chore: ensure pending commits order
Commits would be added to the chain of pending commits based only on the
rebalancing state. Given that the rebalancing state is updated (via
`ConsumerRebalanceListener.onPartitionsAssigned`) separate from when
pending commits are processed (after `poll`), it could happen that
commits emitted later would be processed before earlier pending ones.

This updates the condition for queueing commits to take into account the
prior existence of pending commits.

At the moment, this possibility is theoretical as commit operations are
serialized via `KafkaConsumerActor`'s request queue, and doesn't happen
concurrently to polls. That said, the cost of the fix is trivial and
being explicit about the conditions may prevent future bugs, if the
surrounding context changes.