fix: reset spillover after poll loop
In the face of backpressure, spillover records that couldn't be offered
to the queue would be registered, and wouldn't alwayd be cleared
afterwards, leading to duplicate records being offered. This is
addressed by ensuring spillover records are always reset at the end of
the poll loop.

In an earlier iteration of #1369 spillover records were reset early in
the poll loop. At the end of the loop only remaining spillover records
needed to be recorded. In the approach that ended up being submitted
spillover records are reset only once at the end of the poll loop, but
logic from the earlier iteration was inadvertently left behind that
skipped this reset operation in the case there were no remaining
spillover records to register.